+++ lib/sparql_client/update/builder.ex
@@ -16,7 +16,7 @@ defmodule SPARQL.Client.Update.Builder do
       {:ok,
        """
        #{prologue}
-       #{sparql_update_data_keyword(update_form)} DATA {
+       #{update_data_keyword(update_form)} DATA {
        #{triples}
        }
        """}
@@ -46,24 +46,57 @@ defmodule SPARQL.Client.Update.Builder do
     Turtle.write_string(data, Keyword.merge(opts, only: :triples))
   end
 
-  defp sparql_update_data_keyword(:insert_data), do: "INSERT"
-  defp sparql_update_data_keyword(:delete_data), do: "DELETE"
-
   def load(from, to, silent) do
     {:ok, "LOAD " <> silent_fragment(silent) <> iri_ref(from) <> into_graph_fragment(to)}
   end
 
   def clear(graph_iri, silent) do
-    {:ok, "CLEAR " <> silent_fragment(silent) <> clear_graph_fragment(graph_iri)}
+    {:ok, "CLEAR " <> silent_fragment(silent) <> clear_graph_identifier(graph_iri)}
   end
+
+  def create(graph_iri, silent) do
+    {:ok, "CREATE " <> silent_fragment(silent) <> "GRAPH #{iri_ref(graph_iri)}"}
+  end
+
+  def drop(graph_iri, silent) do
+    {:ok, "DROP " <> silent_fragment(silent) <> clear_graph_identifier(graph_iri)}
+  end
+
+  def copy(from, to, silent) do
+    graph_update(:copy, from, to, silent)
+  end
+
+  def move(from, to, silent) do
+    graph_update(:move, from, to, silent)
+  end
+
+  def add(from, to, silent) do
+    graph_update(:add, from, to, silent)
+  end
+
+  def graph_update(update_form, from, to, silent) do
+    {:ok,
+     "#{graph_update_keyword(update_form)} " <>
+       silent_fragment(silent) <> "#{graph_identifier(from)} TO #{graph_identifier(to)}"}
+  end
+
+  defp update_data_keyword(:insert_data), do: "INSERT"
+  defp update_data_keyword(:delete_data), do: "DELETE"
+
+  defp graph_update_keyword(:copy), do: "COPY"
+  defp graph_update_keyword(:move), do: "MOVE"
+  defp graph_update_keyword(:add), do: "ADD"
 
   defp into_graph_fragment(nil), do: ""
   defp into_graph_fragment(iri), do: " INTO GRAPH #{iri_ref(iri)}"
 
-  defp clear_graph_fragment(:default), do: "DEFAULT"
-  defp clear_graph_fragment(:named), do: "NAMED"
-  defp clear_graph_fragment(:all), do: "ALL"
-  defp clear_graph_fragment(iri), do: "GRAPH #{iri_ref(iri)}"
+  defp clear_graph_identifier(:default), do: "DEFAULT"
+  defp clear_graph_identifier(:named), do: "NAMED"
+  defp clear_graph_identifier(:all), do: "ALL"
+  defp clear_graph_identifier(iri), do: "GRAPH #{iri_ref(iri)}"
+
+  defp graph_identifier(:default), do: "DEFAULT"
+  defp graph_identifier(iri), do: "GRAPH #{iri_ref(iri)}"
 
   defp silent_fragment(true), do: "SILENT "
   defp silent_fragment(_), do: ""